add force option to the permission creator service

+++ src/Kunstmaan/AdminBundle/Helper/Creators/ACLPermissionCreator.php
@@ -39,33 +39,37 @@ class ACLPermissionCreator
     /**
      * @param mixed $object
      * @param array $aces
+     * @param bool  $force
      */
-    private function init($object, $aces)
+    private function init($object, $aces, $force = false)
     {
         $aclProvider = $this->aclProvider;
         $strategy = $this->objectIdentityRetrievalStrategy;
 
         $objectIdentity = $strategy->getObjectIdentity($object);
-        try {
-            $aclProvider->deleteAcl($objectIdentity);
-        } catch (AclNotFoundException $e) {
-            // Do nothing
+        if ($force || $aclProvider->findAcl($objectIdentity) == null) {
+            try {
+                $aclProvider->deleteAcl($objectIdentity);
+            } catch (AclNotFoundException $e) {
+                // Do nothing
+            }
+
+            $acl = $aclProvider->createAcl($objectIdentity);
+
+            foreach ($aces as $ace) {
+                $acl->insertObjectAce($ace['identity'], $ace['mask']);
+            }
+
+            $aclProvider->updateAcl($acl);
         }
-
-        $acl = $aclProvider->createAcl($objectIdentity);
-
-        foreach ($aces as $ace) {
-            $acl->insertObjectAce($ace['identity'], $ace['mask']);
-        }
-
-        $aclProvider->updateAcl($acl);
     }
 
     /**
      * @param mixed $object
      * @param mixed $example
+     * @param bool  $force
      */
-    public function initByExample($object, $example)
+    public function initByExample($object, $example, $force = false)
     {
         $aclProvider = $this->aclProvider;
         $strategy = $this->objectIdentityRetrievalStrategy;
@@ -84,7 +88,7 @@ class ACLPermissionCreator
                 );
             }
         }
-        $this->init($object, $aces);
+        $this->init($object, $aces, $force);
     }
 
     /**
@@ -93,8 +97,9 @@ class ACLPermissionCreator
      *        with as key the name of the role you want to set the permissions for
      *        and as value the mask you want to use
      *        for example array('ROLE_GUEST' => MaskBuilder::MASK_EDIT | MaskBuilder::MASK_PUBLISH)
+     * @param bool  $force
      */
-    public function initByMap($object, $map)
+    public function initByMap($object, $map, $force = false)
     {
         $aces = array();
         foreach ($map as $key => $value) {
@@ -104,7 +109,7 @@ class ACLPermissionCreator
             );
         }
 
-        $this->init($object, $aces);
+        $this->init($object, $aces, $force);
     }
 
 }
